Skip passing enable-ha flag to test script (#173)
* skip passing enable-ha flag to test script

Signed-off-by: Paul S. Schweigert <paulschw@us.ibm.com>

Serving uses the `--enable-ha` flag in several of its e2e tests, but
parsing it out causes the flag to end up getting passed twice (once as
part of the script, and once from the parse out). So this PR fixes that
problem by not passing it as part of the e2e script commands.

For more details, see https://github.com/knative/serving/issues/12788

* use run-tests flag to prevent double parsing

Signed-off-by: Paul S. Schweigert <paulschw@us.ibm.com>