Trap calls are now executed in LIFO order (#249)
We tend to add traps at various points in our test scripts.
The assumption we have is the trap will be executed when
external resources are in a certain state.

ie.
1. Setup Cluster - trap delete_cluster
2. Setup Knative - trap teardown_knative
3. Setup Tests   - trap dump_test_failure

Prior traps are executed in FIFO order so the cluster
would be torn down before the test dump would occur.

In LIFO order the test dump would occur first, then the
teardown of knative and then finally the cluster.